# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hariri Foundation API",
        "description": "API documentation for Users Management",
        "contact": {
            "email": "support@haririfoundation.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/project-activities": {
            "get": {
                "tags": [
                    "Project Activities"
                ],
                "summary": "Get all project activities or a specific one by ID",
                "operationId": "f6198524779b58340d976625b0db6704",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Optional Project Activity UUID to fetch a specific record",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project Activities retrieved successfully"
                    },
                    "404": {
                        "description": "Project Activity not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Project Activities"
                ],
                "summary": "Create a new Project Activity",
                "operationId": "9190251e7175e0e94e4efd29b621a4fd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_center_id",
                                    "activity_id"
                                ],
                                "properties": {
                                    "project_center_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "a2f4d8c7-8b5f-4d88-9d7c-12e456789abc"
                                    },
                                    "activity_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "b3f4a3d9-91c6-4899-bc2e-7a6b7b9b91e4"
                                    },
                                    "external_id": {
                                        "type": "string",
                                        "example": "proj_act_2025_10_center1_activityA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Project Activity created successfully"
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                }
            }
        },
        "/api/project-activities/{id}": {
            "put": {
                "tags": [
                    "Project Activities"
                ],
                "summary": "Update a Project Activity",
                "operationId": "fce78e28b11543572e5de259253b5422",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project Activity UUID to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "project_center_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "a2f4d8c7-8b5f-4d88-9d7c-12e456789abc"
                                    },
                                    "activity_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "b3f4a3d9-91c6-4899-bc2e-7a6b7b9b91e4"
                                    },
                                    "external_id": {
                                        "type": "string",
                                        "example": "proj_act_updated"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Project Activity updated successfully"
                    },
                    "404": {
                        "description": "Project Activity not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Project Activities"
                ],
                "summary": "Delete a Project Activity by UUID",
                "operationId": "e918216171614dc4a0324954df961200",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project Activity UUID to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Project Activity deleted successfully"
                    },
                    "404": {
                        "description": "Project Activity not found"
                    }
                }
            }
        },
...
    },
...
}